refactor: fix import path, use package name, fix build script
