+++ package.json
@@ -6,10 +6,11 @@
   "scripts": {
     "dev": "pnpm --filter genui-sdk-playground dev",
     "build:projects": "pnpm --filter tiny-schema-renderer build && pnpm --filter tiny-schema-renderer-ng build",
-    "build:lib": "pnpm --filter schema-card-renderer build",
+    "build:lib": "pnpm --filter @opentiny/genui-sdk-core build && pnpm --filter @opentiny/genui-sdk-vue build",
     "build:plugins": "pnpm --filter vite-commit-hash-plugin build",
-    "prebuild": "pnpm build:projects && pnpm build:lib && pnpm build:plugins && pnpm --filter schema-renderer-ng-adpater build",
-    "build": "pnpm --filter genui-sdk-playground build"
+    "prebuild:playground": "pnpm build:projects && pnpm build:lib && pnpm build:plugins && pnpm --filter schema-renderer-ng-adpater build",
+    "build:playground": "pnpm --filter genui-sdk-playground build",
+    "build": "pnpm build:playground"
   },
   "keywords": [],
   "author": "OpenTiny",
