+++ packages/core/src/index.ts
@@ -1,1 +1,4 @@
+export * from './delta-patcher';
+export * from './protocols';
 export * from './prompt-generator';
+export * from './delta-json-path-selector';
