+++ packages/core/src/prompt-generator/handle-snippets.ts
@@ -1,4 +1,4 @@
-import type { IMaterials } from '../protocols/materials';
+import type { IMaterials } from '../protocols';
 import type { IWhiteList } from './prompt';
 
 export const extractSnippets = (materialsList: any[]) => {
