+++ packages/core/src/prompt-generator/index.ts
@@ -1,1 +1,2 @@
 export * from './prompt';
+export * from './gen-prompt-config';
