+++ packages/core/src/protocols/index.ts
@@ -1,0 +1,4 @@
+export * from './materials';
+export * from './schema';
+export * from './render-config';
+export * from './chat';
