+++ packages/core/src/protocols/render-config.ts
@@ -1,0 +1,6 @@
+export interface IRendererConfig {
+  materialsList: any[];
+  examples: any[];
+  whiteList: string[];
+  wrapperComponent?: string;
+}
