+++ packages/core/vite.config.ts
