+++ packages/frameworks/vue/src/chat/GeneratingComponent.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup>
 import { ref, onMounted, onBeforeUnmount } from 'vue';
-import type { INotificationPayload } from '../../../../core/src/protocols/chat';
+import type { INotificationPayload } from '@opentiny/genui-sdk-core';
 import type { IGeneratingComponentProps } from './chat.types';
 import { useI18n } from './i18n'; //TODO: replace with package name
