+++ sites/playground/server/tsconfig.dev.json
@@ -3,12 +3,6 @@
   "compilerOptions": {
     "allowSyntheticDefaultImports": true,
     "paths": {
-      "@opentiny/genui-sdk": [
-        "../../../packages/genui-renderer/src/index.ts"
-      ],
-      "@opentiny/genui-sdk/server": [
-        "../../../packages/genui-renderer/src/server/index.ts"
-      ],
       "@opentiny/genui-sdk-core": [
         "../../../packages/core/src/index.ts"
       ]
