+++ sites/playground/web/package.json
@@ -11,7 +11,7 @@
       "preview": "vite preview"
     },
     "dependencies": {
-      "@opentiny/genui-sdk": "workspace:*",
+      "@opentiny/genui-sdk-core": "workspace:*",
       "@opentiny/genui-sdk-vue": "workspace:*",
       "@opentiny/tiny-robot": "0.3.1-alpha.3",
       "@opentiny/tiny-robot-kit": "0.3.1-alpha.3",
